moved or deleted mock Applicatrion vertexes
