+++ unittests/model_tests/splitter_tests/test_one_app_one_machine.py
@@ -15,52 +15,22 @@
 
 import unittest
 from pacman.exceptions import PacmanConfigurationException
-from pacman.model.graphs.application import ApplicationVertex
 from pacman.model.graphs.application.abstract import (
     AbstractOneAppOneMachineVertex)
 from pacman.model.partitioner_splitters import SplitterOneAppOneMachine
-
-
-class MockLegacy(ApplicationVertex):
-    def __init__(self, label):
-        super().__init__(
-            label=label, constraints=None,  max_atoms_per_core=None)
-
-    def get_resources_used_by_atoms(self, vertex_slice):
-        pass
-
-    def create_machine_vertex(
-            self, vertex_slice, resources_required, label=None,
-            constraints=None):
-        pass
-
-    def n_atoms(self):
-        pass
-
-
-class MockNot(ApplicationVertex):
-    def __init__(self, label):
-        super().__init__(
-            label=label, constraints=None,  max_atoms_per_core=None)
-
-    def n_atoms(self):
-        pass
-
-
-class MockApi(AbstractOneAppOneMachineVertex):
-    def __init__(self, label):
-        super().__init__(machine_vertex=None, label=label, constraints=None)
+from pacman_test_objects import NonLegacyApplicationVertex
 
 
 class TestSplitterOneAppOneMachine(unittest.TestCase):
     def test_legacy(self):
         splitter = SplitterOneAppOneMachine("foo")
-        v1 = MockNot("v1")
+        v1 = NonLegacyApplicationVertex("v1")
         a = str(splitter)
         self.assertIsNotNone(a)
         with self.assertRaises(PacmanConfigurationException):
             splitter.set_governed_app_vertex(v1)
-        v2 = MockApi("v1")
+        v2 = AbstractOneAppOneMachineVertex(
+            machine_vertex=None, label="v1", constraints=None)
         splitter.set_governed_app_vertex(v2)
         a = str(splitter)
         self.assertIsNotNone(a)
