+++ unittests/model_tests/splitter_tests/test_splitter_slice_legacy.py
@@ -15,52 +15,9 @@
 
 import unittest
 from pacman.exceptions import PacmanConfigurationException
-from pacman.model.graphs.application import ApplicationVertex
-from pacman.model.partitioner_interfaces import LegacyPartitionerAPI
 from pacman.model.partitioner_splitters import SplitterSliceLegacy
-
-
-class MockLegacyApi(ApplicationVertex, LegacyPartitionerAPI):
-    def __init__(self, label):
-        super().__init__(
-            label=label, constraints=None,  max_atoms_per_core=None)
-
-    def get_resources_used_by_atoms(self, vertex_slice):
-        pass
-
-    def create_machine_vertex(
-            self, vertex_slice, resources_required, label=None,
-            constraints=None):
-        pass
-
-    def n_atoms(self):
-        pass
-
-
-class MockLegacy(ApplicationVertex):
-    def __init__(self, label):
-        super().__init__(
-            label=label, constraints=None,  max_atoms_per_core=None)
-
-    def get_resources_used_by_atoms(self, vertex_slice):
-        pass
-
-    def create_machine_vertex(
-            self, vertex_slice, resources_required, label=None,
-            constraints=None):
-        pass
-
-    def n_atoms(self):
-        pass
-
-
-class MockNotLegacy(ApplicationVertex):
-    def __init__(self, label):
-        super().__init__(
-            label=label, constraints=None,  max_atoms_per_core=None)
-
-    def n_atoms(self):
-        pass
+from pacman_test_objects import (
+    DuckLegacyApplicationVertex, NonLegacyApplicationVertex, SimpleTestVertex)
 
 
 class TestSplitterSliceLegacy(unittest.TestCase):
@@ -68,24 +25,24 @@ class TestSplitterSliceLegacy(unittest.TestCase):
         splitter = SplitterSliceLegacy("foo")
         a = str(splitter)
         self.assertIsNotNone(a)
-        v1 = MockLegacyApi("v1")
+        v1 = SimpleTestVertex("v1")
         splitter.set_governed_app_vertex(v1)
         a = str(splitter)
         self.assertIsNotNone(a)
         splitter.set_governed_app_vertex(v1)
-        v2 = MockLegacyApi("v1")
+        v2 = SimpleTestVertex("v1")
         with self.assertRaises(PacmanConfigurationException):
             splitter.set_governed_app_vertex(v2)
 
     def test_not_api(self):
         splitter = SplitterSliceLegacy("foo")
-        v1 = MockNotLegacy("v1")
+        v1 = NonLegacyApplicationVertex("v1")
         with self.assertRaises(PacmanConfigurationException):
             splitter.set_governed_app_vertex(v1)
 
     def test_legacy(self):
         splitter = SplitterSliceLegacy("foo")
-        v1 = MockLegacy("v1")
+        v1 = DuckLegacyApplicationVertex("v1")
         splitter.set_governed_app_vertex(v1)
 
     def test_max_atoms_start_variable(self):
